% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{tsclean2}
\alias{tsclean2}
\title{tsclean2
values as well, using a user-specified function from the imputeTS package.}
\usage{
tsclean2(
  x,
  replace.missing = TRUE,
  iterate = 2,
  lambda = NULL,
  method = na_interpolation,
  ...
)
}
\arguments{
\item{x}{The vector to be cleansed}

\item{replace.missing}{If TRUE, not only replaces outliers, but also replaces missing values}

\item{iterate}{The number of iterations required; pass-through to tsoutliers.}

\item{lambda}{Box-Cox transformation parameter. If lambda="auto", then a transformation
is automatically selected using BoxCox.lambda. The transformation is ignored if NULL.
Otherwise, data transformed before model is estimated. Pass-through to tsoutliers.}

\item{method}{See imputeTS documentation for interpolation methods}

\item{...}{Other parameters to be passed to method}
}
\value{
A vector with outliers replaced by imputed values
}
\description{
tsclean2
values as well, using a user-specified function from the imputeTS package.
}
