% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{History cleansing}
\usage{
clean(x, cutoff = 0.999, randomize = FALSE, method = na_interpolation)
}
\arguments{
\item{x}{The vector to be cleansed}

\item{cutoff}{The cutoff p-value used to identify "too long" a run of zero values}

\item{randomize}{Randomize the replacement values (using a Poisson dist'n)}

\item{method}{The  method used to calculate replacement values}
}
\value{
The vector with values replaced
}
\description{
Identify and replace outliers, including NAs and long runs of zero values,
in an integer-valued time series
}
\examples{
x <- rpois(50,5)
x[31:35] <- 0
z <- clean(x, method = na_kalman, randomize=TRUE)
print(z[31:35])
}
